entities changes and service changes

+++ src/main/java/com/grocio/backend/dto/OrderItemRequestDTO.java
@@ -1,10 +1,12 @@
 package com.grocio.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class OrderItemRequestDTO {
     private Long productId;
     private Integer quantity;
-    private Double priceAtThatTime;
+    private BigDecimal priceAtThatTime;
 }
