+++ src/main/java/com/grocio/backend/dto/ProductDetailsDTO.java
@@ -1,14 +1,16 @@
 package com.grocio.backend.dto;
 
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class ProductDetailsDTO {
     private Long id;
     private String name;
-    private Double price;
-    private Double oldPrice;
+    private BigDecimal price;
+    private BigDecimal oldPrice;
     private String unit;
     private String imageUrl;
     private String badge;
