+++ src/main/java/com/grocio/backend/dto/RelatedProductDTO.java
@@ -1,13 +1,15 @@
 package com.grocio.backend.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
 public class RelatedProductDTO {
     private Long id;
     private String name;
-    private Double price;
-    private Double oldPrice;
+    private BigDecimal price;
+    private BigDecimal oldPrice;
     private String unit;
     private String imageUrl;
 }
